Compute next state for _all_ cells.
Previously the state was computed only for live cells and their
neighbors. But this would fail for rules where a dead cell surrounded by
live cells comes alive (which is half of all randomly-chosen rules).

One could detect such rules by seeing if bit 0 is set in the rule (or,
equivalently, seeing if the rule number is odd), and then use a
different step function in that case, but I think it's probably simpler
to just use the same logic for all rules.